+++ NEWS.md
@@ -1,4 +1,4 @@
-# pedquant 0.2.4.999
+# pedquant 0.2.5
 
 * fixed a bug in pq_plot function to support candle chart
 * the md_stock_symbol supports to query the constituent symbols of Chinese securities index